+++ checks/default.nix
@@ -189,5 +189,27 @@ myPkgs // {
     src = ./workspace-git;
     pname = "workspace-git";
   };
+
+  nextestSimple = myLib.cargoNextest {
+    src = ./simple;
+    pname = "nextest-simple";
+    cargoArtifacts = null;
+  };
+
+  nextestPartitionsCount = myLib.cargoNextest {
+    src = ./simple;
+    pname = "nextest-partitions-count";
+    partitions = 4;
+    partitionType = "count";
+    cargoArtifacts = null;
+  };
+
+  nextestPartitionsHash = myLib.cargoNextest {
+    src = ./simple;
+    pname = "nextest-partitions-hash";
+    partitions = 4;
+    partitionType = "hash";
+    cargoArtifacts = null;
+  };
 })
 )